LayoutTests/accessibility/aria-liveregion-marquee-default.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body id="body">

<div tabindex="0" id="marquee1" role="marquee">marquee1</div>
<div tabindex="0" id="marquee2" role="marquee" aria-live="assertive">marquee2</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that the marquee role has the correct default aria-live status (off).");

    if (window.accessibilityController) {
          // Make sure that regular elements are not exposing ARIA live attributes.
          shouldBe("accessibilityController.accessibleElementById('marquee1').stringAttributeValue('AXARIALive')", "'off'");
          shouldBe("accessibilityController.accessibleElementById('marquee2').stringAttributeValue('AXARIALive')", "'assertive'");
    }

</script>

<script src="../resources/js-test-post.js"></script>
</body>
</html>
