LayoutTests/accessibility/aria-liveregion-marquee-default-expected.txt

marquee1
marquee2
This tests that the marquee role has the correct default aria-live status (off).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.accessibleElementById('marquee1').stringAttributeValue('AXARIALive') is 'off'
PASS accessibilityController.accessibleElementById('marquee2').stringAttributeValue('AXARIALive') is 'assertive'
PASS successfullyParsed is true

TEST COMPLETE
